requirements.txt
requests==2.11.1
requests-toolbelt==0.7.0
moviepy==0.2.2.11

ffmpeg
gpac
python-cups
